pyproject.toml
[project]
name = "pongrl-experiments"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gymnasium[accept-rom-license,atari]>=1.2.3",
    "pillow>=12.0.0",
    "shimmy>=2.0.0",
]
